requirements.txt
tweepy>=4.14.0
anthropic>=0.25.0
crewai>=0.28.0
jinja2>=3.1.0
python-dotenv>=1.0.0
